requirements.txt
chela==1.0.5
matplotlib==3.1.2
mendeleev==0.6.0
numpy==1.18.5
pandas==1.1.5
PyYAML>=3.12
scikit_learn==0.22.1
seaborn==0.11.0
scipy==1.4.1
tensorflow==2.3.0
keras-tuner==1.0.2
jupytext==1.14.1
mlflow== 1.23.1

